Port to .NET 5
