--- CenterTaskbar/App.config
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2"/>
-    </startup>
-    <system.diagnostics>
-        <trace>
-          <listeners>
-          </listeners>
-        </trace>
-      </system.diagnostics>
-</configuration>
